# backends/auth/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /auth/v1/login:
        post:
            tags:
                - Auth
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.auth.v1.LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.LoginResp'
    /auth/v1/sign_up:
        post:
            tags:
                - Auth
            operationId: Auth_CreateAccount
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.auth.v1.CreateAccountReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.CreateAccountResp'
    /auth/version:
        get:
            tags:
                - Version
            operationId: Version_Version
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/version.VersionReply'
components:
    schemas:
        api.auth.v1.CreateAccountReq:
            type: object
            properties:
                username:
                    type: string
                pwd:
                    type: string
        api.auth.v1.CreateAccountResp:
            type: object
            properties:
                uid:
                    type: string
        api.auth.v1.LoginReq:
            type: object
            properties:
                uid:
                    type: string
                pwd:
                    type: string
        api.auth.v1.LoginResp:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
        version.VersionReply:
            type: object
            properties:
                name:
                    type: string
                version:
                    type: string
tags:
    - name: Auth
    - name: Version